chore: move model to package
